Rename packet.py to packets.py, add different packet types
